+++ packages/backend/src/projectAdapters/dbtBaseProjectAdapter.ts
@@ -29,7 +29,7 @@ import {
     type LightdashProjectConfig,
 } from '@lightdash/common';
 import { WarehouseClient } from '@lightdash/warehouses';
-import { promises as fs } from 'fs';
+import fs from 'fs/promises';
 import path from 'path';
 import Logger from '../logging/logger';
 import { CachedWarehouse, DbtClient, ProjectAdapter } from '../types';