+++ packages/cli/src/handlers/compile.ts
@@ -2,18 +2,15 @@ import {
     attachTypesToModels,
     convertExplores,
     DbtManifestVersion,
-    DEFAULT_SPOTLIGHT_CONFIG,
     getCompiledModels,
     getDbtManifestVersion,
     getModelsFromManifest,
     getSchemaStructureFromDbtModels,
     isExploreError,
     isSupportedDbtAdapter,
-    loadLightdashProjectConfig,
     ParseError,
     WarehouseCatalog,
 } from '@lightdash/common';
-import { promises as fs } from 'fs';
 import path from 'path';
 import { v4 as uuidv4 } from 'uuid';
 import { LightdashAnalytics } from '../analytics/analytics';
@@ -21,6 +18,7 @@ import { getDbtContext } from '../dbt/context';
 import { loadManifest } from '../dbt/manifest';
 import { validateDbtModel } from '../dbt/validation';
 import GlobalState from '../globalState';
+import { readAndLoadLightdashProjectConfig } from '../lightdash-config';
 import * as styles from '../styles';
 import { DbtCompileOptions, maybeCompileModelsAndJoins } from './dbt/compile';
 import { getDbtVersion } from './dbt/getDbtVersion';
@@ -34,25 +32,6 @@ export type CompileHandlerOptions = DbtCompileOptions & {
     vars: string | undefined;
     verbose: boolean;
     startOfWeek?: number;
-};
-
-const readAndLoadLightdashProjectConfig = async (projectDir: string) => {
-    const configPath = path.join(projectDir, 'lightdash.config.yml');
-    try {
-        const fileContents = await fs.readFile(configPath, 'utf8');
-        const config = await loadLightdashProjectConfig(fileContents);
-        return config;
-    } catch (e) {
-        GlobalState.debug(`No lightdash.config.yml found in ${configPath}`);
-
-        if (e instanceof Error && 'code' in e && e.code === 'ENOENT') {
-            // Return default config if file doesn't exist
-            return {
-                spotlight: DEFAULT_SPOTLIGHT_CONFIG,
-            };
-        }
-        throw e;
-    }
 };
 
 export const compile = async (options: CompileHandlerOptions) => {